Move core/containers/ to core/components/
