+++ src/core/components/error-simulation/SimulateClientError/styles.scss
