+++ src/core/components/error-simulation/SimulateSyncError/index.js
