+++ tests/unit/core/components/error-simulation/TestSimulateSyncError.js
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { renderIntoDocument } from 'react-dom/test-utils';
 
-import { SimulateSyncErrorBase } from 'core/containers/error-simulation/SimulateSyncError';
+import { SimulateSyncErrorBase } from 'core/components/error-simulation/SimulateSyncError';
 
-describe('SimulateSyncError', () => {
+describe(__filename, () => {
   function render(props = {}) {
     return renderIntoDocument(<SimulateSyncErrorBase {...props} />);
   }
